requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
python-docx
sqlalchemy
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0